# perf(symgo): Massive performance optimization - 95% faster, 99.9% less memory