--- a/symgo/object/object.go
+++ b/symgo/object/object.go
@@ -9,6 +9,7 @@ import (
 	"os"
 	"strconv"
 	"strings"
+	"sync"
 
 	"github.com/podhmo/go-scan/scanner"
 )
@@ -101,6 +102,12 @@ func (s *String) Type() ObjectType { return STRING_OBJ }
 // Inspect returns a string representation of the String's value.
 func (s *String) Inspect() string { return fmt.Sprintf("%q", s.Value) }
 
+// Release returns the String object to the pool.
+func (s *String) Release() {
+	s.Value = ""
+	stringPool.Put(s)
+}
+
 // --- Integer Object ---
 
 // Integer represents an integer value.
@@ -115,6 +122,12 @@ func (i *Integer) Type() ObjectType { return INTEGER_OBJ }
 // Inspect returns a string representation of the Integer's value.
 func (i *Integer) Inspect() string { return strconv.FormatInt(i.Value, 10) }
 
+// Release returns the Integer object to the pool.
+func (i *Integer) Release() {
+	i.Value = 0
+	integerPool.Put(i)
+}
+
 // --- Float Object ---
 
 // Float represents a float value.
@@ -128,6 +141,12 @@ func (f *Float) Type() ObjectType { return FLOAT_OBJ }
 
 // Inspect returns a string representation of the Float's value.
 func (f *Float) Inspect() string { return strconv.FormatFloat(f.Value, 'f', -1, 64) }
+
+// Release returns the Float object to the pool.
+func (f *Float) Release() {
+	f.Value = 0
+	floatPool.Put(f)
+}
 
 // --- Complex Object ---
 
@@ -162,7 +181,30 @@ var (
 	TRUE = &Boolean{Value: true}
 	// FALSE is the singleton false value.
 	FALSE = &Boolean{Value: false}
+	// NIL is the singleton nil value.
+	NIL = &Nil{}
 )
+
+// NewInteger creates a new Integer object from the pool.
+func NewInteger(value int64) *Integer {
+	obj := integerPool.Get().(*Integer)
+	obj.Value = value
+	return obj
+}
+
+// NewString creates a new String object from the pool.
+func NewString(value string) *String {
+	obj := stringPool.Get().(*String)
+	obj.Value = value
+	return obj
+}
+
+// NewFloat creates a new Float object from the pool.
+func NewFloat(value float64) *Float {
+	obj := floatPool.Get().(*Float)
+	obj.Value = value
+	return obj
+}
 
 // --- Function Object ---
 
@@ -510,10 +552,42 @@ type Environment struct {
 	outer *Environment
 }
 
+// Object pools for reusing common objects
+var (
+	envPool = sync.Pool{
+		New: func() interface{} {
+			return &Environment{
+				store: make(map[string]Object),
+				outer: nil,
+			}
+		},
+	}
+	integerPool = sync.Pool{
+		New: func() interface{} {
+			return &Integer{}
+		},
+	}
+	stringPool = sync.Pool{
+		New: func() interface{} {
+			return &String{}
+		},
+	}
+	floatPool = sync.Pool{
+		New: func() interface{} {
+			return &Float{}
+		},
+	}
+)
+
 // NewEnvironment creates a new, top-level environment.
 func NewEnvironment() *Environment {
-	s := make(map[string]Object)
-	return &Environment{store: s, outer: nil}
+	env := envPool.Get().(*Environment)
+	// Reset the environment state
+	for k := range env.store {
+		delete(env.store, k)
+	}
+	env.outer = nil
+	return env
 }
 
 // NewEnclosedEnvironment creates a new environment that is enclosed by an outer one.
@@ -579,6 +653,15 @@ func (e *Environment) WalkLocal(fn func(name string, obj Object) bool) {
 		if !fn(name, obj) {
 			return
 		}
+	}
+}
+
+// Release returns the environment to the pool for reuse.
+// Only call this on environments that are no longer needed.
+func (e *Environment) Release() {
+	// Only release if this is not an outer environment being used by others
+	if e.outer == nil {
+		envPool.Put(e)
 	}
 }
 
@@ -781,6 +864,4 @@ func (uf *UnresolvedFunction) Inspect() string {
 // --- Global Instances ---
 
 // Pre-create global instances for common values to save allocations.
-var (
-	NIL = &Nil{}
-)
+// (NIL is already defined above)