… docs/summary-symgo-performance-optimization.md